refactor(domain): rename Entity to Aggregate following DDD terminology
